🔨 refactor: 파일 구조 변경 & 임시로 search 권한 개방
