+++ build.gradle
@@ -60,8 +60,7 @@ dependencies {
 
 	// elasticSearch
 	implementation 'org.springframework.boot:spring-boot-starter-data-elasticsearch'
-	implementation 'org.elasticsearch.client:elasticsearch-rest-high-level-client:7.17.18'
-	implementation 'org.elasticsearch:elasticsearch:7.17.18'
+
 
 }
